+++ pyphi/__about__.py
@@ -5,12 +5,12 @@
 """PyPhi metadata."""
 
 __title__ = 'pyphi'
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 __description__ = 'Python library for computing integrated information.'
 __author__ = 'William GP Mayner'
 __author_email__ = 'wmayner@gmail.com'
 __author_website__ = 'http://willmayner.com'
-__copyright__ = 'Copyright 2014-2018 Will Mayner'
+__copyright__ = 'Copyright 2014-2019 William GP Mayner'
 __license__ = 'GNU General Public License v3.0'
 __url__ = 'http://github.com/wmayner/pyphi'