+++ pyphi/constants.py
@@ -8,12 +8,10 @@ Package-wide constants.
 
 import pickle
 
-import joblib
-
-from . import config
-
 #: The threshold below which we consider differences in phi values to be zero.
-EPSILON = 10 ** - config.PRECISION
+EPSILON = None
+# NOTE: This is set dynamically by `conf.py` when PRECISION is changed; see
+# `conf.py` for default value.
 
 #: Label for the filesystem cache backend.
 FILESYSTEM = 'fs'
@@ -25,8 +23,9 @@ DATABASE = 'db'
 PICKLE_PROTOCOL = pickle.HIGHEST_PROTOCOL
 
 #: The joblib ``Memory`` object for persistent caching without a database.
-joblib_memory = joblib.Memory(cachedir=config.FS_CACHE_DIRECTORY,
-                              verbose=config.FS_CACHE_VERBOSITY)
+joblib_memory = None
+# NOTE: This is set dynamically by `conf.py` when PRECISION is changed; see
+# `conf.py` for default value.
 
 #: Node states
 OFF = (0,)